TaskConfig is now a property of the TaskExecutor
This way each plugin can specify their own interface for tasks
