+++ task_processing/plugins/mesos/mesos_executor.py
@@ -1,9 +1,15 @@
 import logging
 import os
 import threading
+import uuid
 
 import mesos.native
 from mesos.interface import mesos_pb2
+from pyrsistent import field
+from pyrsistent import PRecord
+from pyrsistent import PVector
+from pyrsistent import pvector
+from pyrsistent import v
 
 from task_processing.interfaces.task_executor import TaskExecutor
 from task_processing.plugins.mesos.execution_framework import (
@@ -16,7 +22,34 @@ logging.basicConfig(format=FORMAT, level=LEVEL)
 log = logging.getLogger(__name__)
 
 
+class MesosTaskConfig(PRecord):
+    uuid = field(type=uuid.UUID, initial=uuid.uuid4)
+    name = field(type=str, initial="default")
+    image = field(type=str, initial="ubuntu:xenial")
+    cmd = field(type=str, initial="/bin/true")
+    cpus = field(type=float,
+                 initial=0.1,
+                 invariant=lambda c: (c > 0, 'cpus > 0'))
+    mem = field(type=float,
+                initial=32.0,
+                invariant=lambda m: (m >= 32, 'mem is >= 32'))
+    disk = field(type=float,
+                 initial=10.0,
+                 invariant=lambda d: (d > 0, 'disk > 0'))
+    volumes = field(type=PVector, initial=v(), factory=pvector)
+    ports = field(type=PVector, initial=v(), factory=pvector)
+    cap_add = field(type=PVector, initial=v(), factory=pvector)
+    ulimit = field(type=PVector, initial=v(), factory=pvector)
+    # TODO: containerization + containerization_args ?
+    docker_parameters = field(type=PVector, initial=v(), factory=pvector)
+
+    def task_id(self):
+        return "{}.{}".format(self.name, str(self.uuid))
+
+
 class MesosExecutor(TaskExecutor):
+    TASK_CONFIG_INTERFACE = MesosTaskConfig
+
     def __init__(self,
                  authentication_principal='taskproc',
                  credential_secret_file=None,
